+++ drv/rainsensorqty/commands/rainsensorqty_HISTORY.sh
@@ -4,8 +4,8 @@
 # Author: androtto
 # file "rainsensorqty_CHECK.sh"
 # test script for checking rain status using drv_rainsensorqty_rain_sensor_get function
-# Version: 0.2.0a
-# Data: 13/Aug/2019
+# Version: 0.2.5
+# Data: 08/Jan/2020
 
 SCRIPTDIR="$(cd `dirname $0` ; pwd )"
 SCRIPTNAME=${0##*/}
@@ -22,7 +22,38 @@ LOG_OUTPUT_DRV_FILE="$DIR_SCRIPT/log/$LOG_OUTPUT_DRV_FILE"
 . ./init.include.sh
 . ./rainsensor.include.sh
 
-rain_history # update rain history file if not
+if [[ $1 = "-force" ]] ; then
+	if [[ -s $RAINSENSORQTY_HISTORY ]] ; then
+		echo backup $RAINSENSORQTY_HISTORY to ${RAINSENSORQTY_HISTORY}.old$$
+		cp $RAINSENSORQTY_HISTORY ${RAINSENSORQTY_HISTORY}.old$$
+	fi
 
-echo "RAIN HISTORY"
-cat $RAINSENSORQTY_HISTORY | rain_when_amount
+	echo "generate all rain events to $RAINSENSORQTY_HISTORY"
+	if ! rainevents > ${RAINSENSORQTY_HISTORY} ; then
+		echo "WARNING: rainevents function had error"
+	fi
+	shift
+fi
+
+if ! rain_history tmp ; then # update rain history with last rain if not
+	echo "WARNING: rain_history function had error"
+fi
+
+cmd="cat"
+if [[ $# > 0 ]] ; then
+	if (( $1 >= 1 )) ; then
+		echo "processing last $1 lines of $RAINSENSORQTY_HISTORYRAW file"
+		cmd="tail -$1"
+	else
+		echo "argument not recognized - exit"
+		exit 1
+	fi
+fi
+echo -e "\n\n"
+
+if [[ -s $RAINSENSORQTY_HISTORY ]] ; then
+	echo "RAIN HISTORY"
+	cat $RAINSENSORQTY_HISTORY $RAINSENSORQTY_HISTORYTMP | $cmd | rain_when_amount
+else
+	echo "WARNING: no \$RAINSENSORQTY_HISTORY file"
+fi